Added simple id exchange of the nodes when connected.
When the nodes are connected, the id's are exchanged. The id's of the NodeConnection represents the main node it is connected to. This simple implementation is only for debugging purposes, no security or checking is involved. When you want to securely know the real id of the other node, you are able to implement your own methods and protocols.